+++ src/main/webui/src/justifications/justifications.table.tsx
@@ -1,7 +1,6 @@
 import {ReactElement} from "react";
 import {JustificationKinds} from "./JustificationsPanel.tsx";
-import {useForm, UseFormReturnType} from "@mantine/form";
-import {Justification, TextFormats} from "../generated/proposalToolSchemas.ts";
+import {Justification} from "../generated/proposalToolSchemas.ts";
 import {Table} from "@mantine/core";
 import JustificationsEditModal from "./edit.modal.tsx";
 
@@ -10,40 +9,12 @@ export type Which = 'scientific' | 'technical';
 
 export type JustificationProps = {
     which : Which,
-    form: UseFormReturnType<Justification>
+    justification: Justification,
     closeModal?: () => void
 }
 
 export default function JustificationsTable(justifications: JustificationKinds)
     : ReactElement {
-
-    const DEFAULT_JUSTIFICATION : Justification = {text: "", format: "ASCIIDOC" };
-
-    const scientificForm: UseFormReturnType<Justification> =
-        useForm<Justification>({
-            initialValues: justifications.scientific ?? DEFAULT_JUSTIFICATION ,
-            validate: {
-                text: (value: string | undefined) =>
-                    (value === "" || value === undefined ?
-                        "Text cannot be empty for a Scientific Justification" : null),
-                format: (value: TextFormats | undefined ) =>
-                    (value !== "LATEX" && value !== "RST" && value !== "ASCIIDOC" ?
-                        'Text format one of: LATEX, RST, or ASCIIDOC' : null)
-            }
-        });
-
-    const technicalForm: UseFormReturnType<Justification> =
-        useForm<Justification>({
-            initialValues: justifications.technical ?? DEFAULT_JUSTIFICATION ,
-            validate: {
-                text: (value: string | undefined) =>
-                    (value === "" || value === undefined ?
-                        "Text cannot be empty for a Technical Justification" : null),
-                format: (value: TextFormats | undefined ) =>
-                    (value !== "LATEX" && value !== "RST" && value !== "ASCIIDOC" ?
-                        'Text format one of: LATEX, RST, or ASCIIDOC' : null)
-            }
-        });
 
     const justificationsHeader = () : ReactElement => {
         return (
@@ -66,17 +37,17 @@ export default function JustificationsTable(justifications: JustificationKinds)
                 <Table.Tr>
                     <Table.Td>Scientific</Table.Td>
                     <Table.Td>???</Table.Td>
-                    <Table.Td>{scientificForm.values.format}</Table.Td>
+                    <Table.Td>{justifications.scientific.format}</Table.Td>
                     <Table.Td>
-                        <JustificationsEditModal which={'scientific'} form={scientificForm}/>
+                        <JustificationsEditModal which={'scientific'} justification={justifications.scientific}/>
                     </Table.Td>
                 </Table.Tr>
                 <Table.Tr>
                     <Table.Td>Technical</Table.Td>
                     <Table.Td>???</Table.Td>
-                    <Table.Td>{technicalForm.values.format}</Table.Td>
+                    <Table.Td>{justifications.technical.format}</Table.Td>
                     <Table.Td>
-                        <JustificationsEditModal which={'technical'} form={technicalForm}/>
+                        <JustificationsEditModal which={'technical'} justification={justifications.technical}/>
                     </Table.Td>
                 </Table.Tr>
             </Table.Tbody>
